!!WIP!! : [physx][gfx] Trying to add moving sands.
This commit is trying to add moving sands to the desert.

Two techniques have been tested :
- First one tries to use RenderTexture and apply vertex deformation in shader. This is not working great as the collider is not updated.
- Second one is moving the mesh vertices and then updating the meshcollider. It seems to be OK performance wise but I am not yet fully convinced... It works great except for the sands rendering it is still yet to be done.

[MISC]
Adding headers gameobject in the scene hierarchy to clean up a bit the gameobject

[CLEANUP]
Moving everything related to ground infos into a new scripts Ground.
It it then doable to get back the ground information, such as friction, depth, velocity and type.